+++ Sources/MapboxNavigation/Feature.swift
@@ -1,0 +1,21 @@
+import Foundation
+import Turf
+
+extension Feature {
+
+    var featureIdentifier: Int64? {
+        guard let featureIdentifier = self.identifier else { return nil }
+        
+        switch featureIdentifier {
+        case .string(let identifier):
+            return Int64(identifier)
+        case .number(let identifier):
+            switch identifier {
+            case .double(let identifier):
+                return Int64(identifier)
+            case .int(let identifier):
+                return Int64(identifier)
+            }
+        }
+    }
+}